+++ R/plotBarcodeBubble.R
@@ -14,10 +14,12 @@
 #' @param orderBarcodes Order barcodes alphanumerical (boolean). For SPLINTR that represents abundance in  original barcode library. Default `TRUE`.
 #'
 #' @return Returns a bubbleplot of barcodes represented by proportion of total pool
+#' @importFrom magrittr "%>%"
+#' @importFrom rlang .data
 #' @export
 #' @examples
 #' data(test.dge)
-#' plotBarcodeBubble(test.dge$counts, proportionCutoff = 10)
+#' plotBarcodeBubble(test.dge, proportionCutoff = 10)
 
 plotBarcodeBubble <- function(dgeObject,
                               title = "Proportional Bubble Plot",
@@ -66,12 +68,10 @@ plotBarcodeBubble <- function(dgeObject,
     as.factor(seq(1, length(rownames(counts))))
   barcodes.proportional$Barcode <- rownames(barcodes.proportional)
 
-  # high prop barcodes
-  # TODO
-  Highbarcodes <-
-    dplyr::filter_all(barcodes.proportional[, 1:(ncol(barcodes.proportional) - 2)],
-                      dplyr::any_vars(. > proportionCutoff))
-
+  # identify all barcodes to label
+  Highbarcodes <- barcodes.proportional %>%
+    dplyr::select(-c(.data$Barcode, .data$Position)) %>%
+    dplyr::filter(dplyr::if_any(tidyselect::where(is.numeric), ~ .x > proportionCutoff))
 
   if (colorDominant) {
     # make all barcodes grey and only color those that are above threshold cutoff
@@ -148,10 +148,10 @@ plotBarcodeBubble <- function(dgeObject,
   bubble.plot <- ggplot2::ggplot(
     barcodes.proportional.melted,
     ggplot2::aes(
-      x = Position,
-      y = Sample,
-      size = Proportion,
-      color = Color
+      x = .data$Position,
+      y = .data$Sample,
+      size = .data$Proportion,
+      color = .data$Color
     )
   ) +
     ggplot2::geom_point(stat = "identity",
@@ -200,7 +200,7 @@ plotBarcodeBubble <- function(dgeObject,
   # facet plot if grouping is provided
   if (!is.null(group)) {
     bubble.plot <- bubble.plot +
-      facet_grid(barcodes.proportional.melted[[group]] ~ .,
+      ggplot2::facet_grid(barcodes.proportional.melted[[group]] ~ .,
                  scales = "free",
                  space = "free")
   }
@@ -208,7 +208,7 @@ plotBarcodeBubble <- function(dgeObject,
   # remove legend
   if (legend == FALSE) {
     bubble.plot <- bubble.plot +
-      theme(legend.position = "None")
+      ggplot2::theme(legend.position = "None")
   }
   return(bubble.plot)
 }