+++ R/plotBarcodeCumSum.R
@@ -19,7 +19,7 @@
 #'
 #' @examples
 #' data(test.dge)
-#' plotBarcodeCumSum(counts = test.dge, sample1 = 'T0-1', samples = c('T0-1','S9-1', 'S8-1'))
+#' plotBarcodeCumSum(test.dge, referenceSample = 'T0-1', samples = c('T0-1','S9-1', 'S8-1'))
 plotBarcodeCumSum <-
   function(dgeObject,
            referenceSample = NULL,