+++ R/plotBarcodeDistance.R
@@ -10,11 +10,12 @@
 #'
 #' @return Returns a heatmap of sample distances using desired clustering
 #'
+#' @importFrom rlang .data
 #' @export
 #'
 #' @examples
 #' data(test.dge)
-#' plotBarcodeDistance(test.dge$counts)
+#' plotBarcodeDistance(test.dge)
 
 plotBarcodeDistance <- function(dgeObject,
                                 method = "euclidean",
@@ -46,7 +47,7 @@ plotBarcodeDistance <- function(dgeObject,
   # Create a ggheatmap
   title <- paste(title, "-", method)
   ggheatmap <-
-    ggplot2::ggplot(melted_dist, ggplot2::aes(`Var2`, `Var1`, fill = `value`)) +
+    ggplot2::ggplot(melted_dist, ggplot2::aes(.data$Var2, .data$Var1, fill = .data$value)) +
     ggplot2::geom_tile(color = "white") +
     ggplot2::scale_fill_viridis_c(option = "magma") +
     ggplot2::theme_bw() +