+++ NEWS.md
@@ -1,3 +1,10 @@
+# bartools 1.0.0
+
+- Updated function parameter inputs across whole package
+- Updated quickstart vignette, added single-cell vignette
+- Added functions aggregateBarcodes, filterBarcodes, plotUmiFilterThresholds, plotUmiPerBarcode, readBartabCounts, plotBarcodesPerCell for single-cell data and plotSampleCumSum for bulk data
+- removed deprecated plotLibraryCumSum, plotLibraryDiversity, proportionalBubbleplot
+
 # bartools 0.2.5
 
 -   Added single cell QC plotting functions plotCellsPerGroup, plotMetrics, plotClusterEnrichment, plotCellsinClusters